wilder.nvim: revert option - disable apply_incsearch_fix
Enabling this option causes the substring matched part of query to be
wrong. Disabling it has merits and drawbacks too: matching is correct
and live, but already highlighted text flickers a bit while typing
search command. Anyway, we do not need this option for incsearch
highlight to work. IT WORKS!